# src/main/resources/META-INF/resources/schema/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "assoparenteleve API",
    "version" : "1.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ],
...
  "paths" : {
...
    "/adulte/{id}" : {
      "get" : {
        "tags" : [ "Adulte" ],
        "summary" : "Get an adulte by ID",
        "description" : "Returns a single adulte",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Adulte"
                }
              }
            }
          },
          "404" : {
            "description" : "Adulte not found"
          }
        }
      }
    },
...
    "/commande" : {
      "put" : {
        "tags" : [ "Commande" ],
        "summary" : "Update an commande",
        "description" : "Returns the updated commande",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Commande"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Commande updated successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Commande"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input"
          }
        }
      },
      "post" : {
        "tags" : [ "Commande" ],
        "summary" : "Create a new commande",
        "description" : "Returns the created commande",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Commande"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Commande created successfully"
          },
          "400" : {
            "description" : "Invalid input"
          }
        }
      }
    },
    "/commande/adulte/{id}" : {
      "get" : {
        "tags" : [ "Commande" ],
        "summary" : "Get an commande by ID",
        "description" : "Returns a single commande",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Commande"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Commande not found"
          }
        }
      }
    },
    "/commande/evenement/{id}" : {
      "get" : {
        "tags" : [ "Commande" ],
        "summary" : "Get an commande by ID",
        "description" : "Returns a single commande",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Commande"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Commande not found"
          }
        }
      }
    },
    "/commande/send-reminder/{id}" : {
      "post" : {
        "tags" : [ "Commande" ],
        "summary" : "Send a reminder email for unpaid orders",
        "description" : "Sends a reminder email to all unpaid orders related to a specific event",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Emails sent successfully"
          },
          "400" : {
            "description" : "Invalid input"
          }
        }
      }
    },
    "/commande/{id}" : {
      "get" : {
        "tags" : [ "Commande" ],
        "summary" : "Get an commande by ID",
        "description" : "Returns a single commande",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Commande"
                }
              }
            }
          },
          "404" : {
            "description" : "Commande not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "Commande" ],
        "summary" : "Delete a commande",
        "description" : "Deletes a commande by ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "commande deleted successfully"
          },
          "404" : {
            "description" : "commande not found"
          }
        }
      }
    },
    "/eleve" : {
      "get" : {
        "tags" : [ "Eleve" ],
        "summary" : "Get all eleves",
        "description" : "Returns a list of all eleves",
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Eleve"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "No eleves found"
          }
        }
      },
      "post" : {
        "tags" : [ "Eleve" ],
        "summary" : "Create a new eleve",
        "description" : "Returns the created eleve",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Eleve"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Eleve created successfully"
          },
          "400" : {
            "description" : "Invalid input"
          }
        }
      }
    },
    "/eleve/{id}" : {
      "get" : {
        "tags" : [ "Eleve" ],
        "summary" : "Get an eleve by ID",
        "description" : "Returns a single eleve",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Eleve"
                }
              }
            }
          },
          "404" : {
            "description" : "Eleve not found"
          }
        }
      }
    },
...
    "/evenement/{id}" : {
      "get" : {
        "tags" : [ "Evenement" ],
        "summary" : "Get an evenement by ID",
        "description" : "Returns a single evenement",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Evenement"
                }
              }
            }
          },
          "404" : {
            "description" : "Evenement not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "Evenement" ],
        "summary" : "Delete an evenement",
        "description" : "Deletes an evenement by ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Evenement deleted successfully"
          },
          "404" : {
            "description" : "Evenement not found"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "Adulte" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "nom" : {
            "type" : "string"
          },
          "prenom" : {
            "type" : "string"
          },
          "mail" : {
            "type" : "string"
          },
          "role" : {
            "$ref" : "#/components/schemas/Role"
          },
          "motDePasse" : {
            "type" : "string"
          },
          "telephone" : {
            "type" : "string"
          },
          "eleves" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Eleve"
            }
          },
          "active" : {
            "type" : "boolean"
          }
        }
      },
      "Commande" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "payer" : {
            "type" : "boolean"
          },
          "distribuer" : {
            "type" : "boolean"
          },
          "evenement" : {
            "format" : "int32",
            "type" : "integer"
          },
          "adulte" : {
            "$ref" : "#/components/schemas/Adulte"
          },
          "eleve" : {
            "$ref" : "#/components/schemas/Eleve"
          },
          "commandeProduits" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CommandeProduit"
            }
          }
        }
      },
      "CommandeProduit" : {
        "type" : "object",
        "properties" : {
          "produit" : {
            "$ref" : "#/components/schemas/Produit"
          },
          "quantite" : {
            "format" : "int32",
            "type" : "integer"
          }
        }
      },
      "Distribution" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "distributeur" : {
            "type" : "string"
          },
          "lieu" : {
            "type" : "string"
          },
          "dateDistribution" : {
            "type" : "string"
          }
        }
      },
      "Eleve" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "nom" : {
            "type" : "string"
          },
          "prenom" : {
            "type" : "string"
          },
          "dateNaissance" : {
            "type" : "string"
          },
          "niveau" : {
            "type" : "string"
          }
        }
      },
      "Evenement" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "nom" : {
            "type" : "string"
          },
          "dateDebut" : {
            "type" : "string"
          },
          "dateFin" : {
            "type" : "string"
          },
          "datePaiement" : {
            "type" : "string"
          },
          "distributionList" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Distribution"
            }
          },
          "produitList" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Produit"
            }
          },
          "valider" : {
            "type" : "boolean"
          },
          "avecEleve" : {
            "type" : "boolean"
          }
        }
      },
      "Produit" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "nom" : {
            "type" : "string"
          },
          "fournisseur" : {
            "type" : "string"
          },
          "prix" : {
            "format" : "double",
            "type" : "number"
          },
          "option" : {
            "type" : "boolean"
          }
        }
      },
      "Role" : {
        "enum" : [ "SUPER_ADMIN", "ADMIN", "PARENT", "PROFESSOR" ],
        "type" : "string"
      }
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "http",
        "description" : "Authentication",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}